implement rename suggestions
