+++ DSharpPlus.Commands/ArgumentModifiers/FromCode/FromCodeAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace DSharpPlus.Commands.ParameterModifiers;
+namespace DSharpPlus.Commands.ArgumentModifiers;
 
 /// <summary>
 /// Removes the need to manually parse code blocks from a string.
