+++ DSharpPlus.Commands/Trees/CommandParameterBuilder.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using DSharpPlus.Commands;
-using DSharpPlus.Commands.Processors.SlashCommands.Modifiers;
+using DSharpPlus.Commands.Processors.SlashCommands.ArgumentModifiers;
 using DSharpPlus.Entities;
 
 public class CommandParameterBuilder
