+++ DSharpPlus.Tests/Commands/Cases/Commands/TestTopLevelCommands.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using DSharpPlus.Commands;
-using DSharpPlus.Commands.ParameterModifiers;
+using DSharpPlus.Commands.ArgumentModifiers;
 using DSharpPlus.Entities;
 
 namespace DSharpPlus.Tests.Commands.Cases.Commands;
